move tests to contours mod, rename variables, fix docs
